pyproject.toml
[build-system]
requires = ["hatchling>=1.27.0"]
build-backend = "hatchling.build"

[project]
name = "photo-critique-agent"
version = "0.1.0"
description = "Local-first photo critique agent for structured image review workflows."
readme = "README.md"
requires-python = ">=3.9"
license = { text = "MIT" }
authors = [
  { name = "Jay" }
]
dependencies = [
  "jinja2>=3.1.4",
  "pydantic>=2.8.0",
  "pyyaml>=6.0.2",
  "typer>=0.12.3",
]

[project.optional-dependencies]
dev = [
  "pytest>=8.3.2",
]

[project.scripts]
photo-critique = "photo_critique_agent.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/photo_critique_agent"]

[tool.hatch.build]
include = [
  "src/photo_critique_agent/personas/*.yaml",
  "src/photo_critique_agent/templates/*.j2",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
pythonpath = ["src"]
